% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write.R
\name{change_file_ext}
\alias{change_file_ext}
\title{change file extension to new ext}
\usage{
change_file_ext(file, ext)
}
\arguments{
\item{file}{character: current file path with old ext}

\item{ext}{character: new ext}
}
\value{
character file path
}
\description{
change file extension to new ext
}
\keyword{internal}
