% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write.R
\name{check_complex_data}
\alias{check_complex_data}
\title{Check whther object can be saved in tabular formats like fst}
\usage{
check_complex_data(x, ext = NULL)
}
\arguments{
\item{x}{object to confirm if its in tabular form}

\item{ext}{Extension of file to be saved. If NULL it only checks whther or not
it is complex data. If character, it checks whther the data could be saved
in that format.}
}
\value{
logical for complex data
}
\description{
Check whther object can be saved in tabular formats like fst
}
\examples{
\dontrun{
# False
check_complex_data(data.frame())

# TRUE
check_complex_data(list())
}
}
\keyword{internal}
