% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file}
\alias{check_file}
\title{Check that file format is supported and that package is available}
\usage{
check_file(file)
}
\arguments{
\item{file}{character: file path to be read}
}
\value{
invisible TRUE
}
\description{
Check that file format is supported and that package is available
}
\keyword{internal}
