% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{confirm_data}
\alias{confirm_data}
\title{Confirms if data has changed}
\usage{
confirm_data(ss, sh, verbose = getOption("stamp.verbose"))
}
\arguments{
\item{ss}{stamps from stamp}

\item{sh}{stamps from hash}

\item{verbose}{logical: whether to display additional information}
}
\value{
invisible TRUE if confirm passes or FALSE otherwise.
}
\description{
Confirms if data has changed
}
\keyword{internal}
