% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{confirm_names}
\alias{confirm_names}
\title{confirm names of objects match}
\usage{
confirm_names(ss, sh)
}
\arguments{
\item{ss}{stamps from stamp}

\item{sh}{stamps from hash}
}
\value{
invisible TRUE if confirm passes or FALSE otherwise.
}
\description{
confirm names of objects match
}
\keyword{internal}
