% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_write.R
\name{deal_with_file_path}
\alias{deal_with_file_path}
\title{Make sure file names and directory paths are working fine}
\usage{
deal_with_file_path(file, ext, create_dir)
}
\arguments{
\item{file}{character: File or connection to write to}

\item{ext}{character: format or extension of file. Default is
`fs::path_ext(file)`}
}
\value{
character vector with file path
}
\description{
Make sure file names and directory paths are working fine
}
\keyword{internal}
