% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.st_alias_register}
\alias{.st_alias_register}
\title{Register or update an alias configuration (internal)}
\usage{
.st_alias_register(alias, root, state_dir, stamp_path)
}
\description{
Register or update an alias configuration (internal)
}
\keyword{internal}
