% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.st_data_folder}
\alias{.st_data_folder}
\title{Get the absolute path to the data folder (internal)}
\usage{
.st_data_folder(alias = NULL)
}
\arguments{
\item{alias}{Optional alias; if NULL, uses "default"}
}
\value{
Character scalar absolute path to the data folder.
}
\description{
Returns the absolute path to the data folder for the given alias.
The data folder name is configurable via st_opts(data_folder = "...").
}
\keyword{internal}
