% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.st_detect_alias_from_path}
\alias{.st_detect_alias_from_path}
\title{Detect which alias a path belongs to (internal)}
\usage{
.st_detect_alias_from_path(path)
}
\description{
Detect which alias a path belongs to (internal)
}
\keyword{internal}
