% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.st_dir_create}
\alias{.st_dir_create}
\title{Ensure directory exists}
\usage{
.st_dir_create(path)
}
\arguments{
\item{path}{Character scalar path to create.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Create \code{path} if it does not already exist. Uses the \code{fs} package
and will create intermediate directories when required.
}
\keyword{internal}
