% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_registry.R
\docType{data}
\name{.st_formats_env}
\alias{.st_formats_env}
\title{Internal formats registry}
\format{
An object of class \code{environment} of length 5.
}
\usage{
.st_formats_env
}
\description{
An internal environment that stores available read/write handlers for
different file formats. Each registry entry is a list with two
elements: \verb{read = function(path, ...)} and \verb{write = function(x, path, ...)}. Users can extend or override formats with
\code{st_register_format()}.
}
\keyword{internal}
