% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{.st_parse_utc_times}
\alias{.st_parse_utc_times}
\title{Parse UTC timestamp strings (internal)}
\usage{
.st_parse_utc_times(chr)
}
\description{
Parse UTC timestamp strings (internal)
}
\keyword{internal}
