% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.st_path_matches_alias}
\alias{.st_path_matches_alias}
\title{Check if a path belongs to an alias's root (internal)}
\usage{
.st_path_matches_alias(path, alias)
}
\description{
Check if a path belongs to an alias's root (internal)
}
\keyword{internal}
