% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{.st_policy_keep_ids}
\alias{.st_policy_keep_ids}
\title{Compute version IDs to keep under a retention policy (internal)}
\usage{
.st_policy_keep_ids(vtab, pol)
}
\arguments{
\item{vtab}{data.frame of versions \strong{sorted newest → oldest}}

\item{pol}{normalized policy from \code{.st_normalize_policy()}}
}
\description{
Compute version IDs to keep under a retention policy (internal)
}
\keyword{internal}
