% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.st_resolve_and_normalize}
\alias{.st_resolve_and_normalize}
\title{Resolve file path and create st_path object (internal)}
\usage{
.st_resolve_and_normalize(file, format = NULL, alias = NULL, verbose = TRUE)
}
\arguments{
\item{file}{character path or st_path object}

\item{format}{optional format override}

\item{alias}{character alias or NULL}

\item{verbose}{logical; if TRUE, emit warnings}
}
\value{
list(sp = st_path object, resolved_path, alias_used, was_bare, rel_path)
}
\description{
Resolve file path and create st_path object (internal)
}
\keyword{internal}
