% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.st_resolve_file_path}
\alias{.st_resolve_file_path}
\title{Resolve file path using alias (internal)}
\usage{
.st_resolve_file_path(file, alias = NULL, verbose = TRUE)
}
\arguments{
\item{file}{character path (bare filename or path with directory)}

\item{alias}{character alias or NULL}

\item{verbose}{logical; if TRUE, emit warnings}
}
\value{
list(path = resolved_path, alias_used = alias_name, was_bare = logical, rel_path = relative_path_from_root)
}
\description{
Resolve file path using alias (internal)
}
\keyword{internal}
