% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_store.R
\name{.st_root_dir}
\alias{.st_root_dir}
\title{Project root directory recorded by st_init() (internal)}
\usage{
.st_root_dir(alias = NULL)
}
\value{
Character scalar absolute path to project root.
}
\description{
Return the absolute project root directory previously recorded by
\code{st_init()}. If not set, defaults to the current working directory.
}
\keyword{internal}
