% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.stamp_state}
\alias{.stamp_state}
\title{Package internal state environment}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.stamp_state
}
\description{
An internal environment used by the package to store lightweight
runtime state (for example the path to the state directory). It is
implemented with an \code{rlang} environment so values can be read and
updated from multiple functions without exporting mutable global
variables.
}
\details{
This object is internal to the package and not intended for end-user
consumption.
}
\keyword{internal}
