% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write.R
\name{ensure_file_path}
\alias{ensure_file_path}
\title{Make sure file names and directory paths are working fine}
\usage{
ensure_file_path(file, recurse)
}
\arguments{
\item{file}{character: File or connection to write to}

\item{recurse}{logical: is \code{TRUE} if directory in \code{file} does not it will
be created. Default is FALSE}
}
\value{
character vector with file path
}
\description{
Make sure file names and directory paths are working fine
}
\keyword{internal}
