% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExample_SC11.R
\docType{data}
\name{Example_SC11}
\alias{Example_SC11}
\title{Simulated Dataset SC11}
\format{An object of class \code{list} of length 4.}
\usage{
data(Example_SC11)
}
\description{
Example dataset from the simulation study in the paper.
}
\examples{
data(Example_SC11)
# run mixure model with quadratic statistic
\donttest{mixtureQuad(1000,Example_SC11$ListBeta,Example_SC11$ListSE,Example_SC11$ListCor,Example_SC11$MafList) }
}
\references{
Andriy Derkach, Ruth M. Pfeiffer (2018)
(\href{https://arxiv.org/abs/1709.06702}{Arxiv})
}
\keyword{datasets}
