% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExample_SC6.R
\docType{data}
\name{Example_SC6}
\alias{Example_SC6}
\title{Simulated Dataset SC6}
\format{An object of class \code{list} of length 4.}
\usage{
data(Example_SC6)
}
\description{
Example dataset from the simulation study in the paper.
}
\examples{
data(Example_SC6)
\donttest{mixtureQuad(1000,Example_SC6$ListBeta,Example_SC6$ListSE,Example_SC6$ListCor,Example_SC6$MafList) # run mixure model with quadratic statistic}
}
\references{
Andriy Derkach, Ruth M. Pfeiffer (2018)
(\href{https://arxiv.org/abs/1709.06702}{Arxiv})
}
\keyword{datasets}
