% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_st_dir}
\alias{format_st_dir}
\title{Format Stamp directory}
\usage{
format_st_dir(st_dir = NULL)
}
\arguments{
\item{st_dir}{character: parent directory to store stamp file (see details).}
}
\value{
formatted directory
}
\description{
Format Stamp directory
}
\keyword{internal}
