% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_st_file}
\alias{format_st_file}
\title{Format stamp file}
\usage{
format_st_file(
  st_dir = NULL,
  st_name = NULL,
  st_ext = getOption("stamp.default.ext"),
  seed = NULL
)
}
\arguments{
\item{st_dir}{character: parent directory to store stamp file (see details).}

\item{st_name}{character: name of stamp in file. All stamp files are prefixed
with value in option "stamp.stamp_prefix", which by default is "st_".
You don't need to add the prefix.}

\item{st_ext}{character: format of stamp file to save. Default is value in
option "stamp.default.ext"}

\item{seed}{numeric: seed for random name. Default is NULL so each time the
the random name generated is the same. Use only for replicability purposes}
}
\value{
formatted directory
}
\description{
Format stamp file
}
\keyword{internal}
