% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_st_name}
\alias{format_st_name}
\title{Format Stamp name}
\usage{
format_st_name(st_name = rand_name())
}
\arguments{
\item{st_name}{}
}
\value{
formatted directory
}
\description{
Format Stamp name
}
\keyword{internal}
