% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read.R
\name{get_reading_fun}
\alias{get_reading_fun}
\title{Reading function depending on format selected}
\usage{
get_reading_fun(ext = "Rds")
}
\arguments{
\item{ext}{character: format or extension of file. Default is
\code{fs::path_ext(file)}}
}
\value{
reading function according to \code{ext}
}
\description{
Reading function depending on format selected
}
\examples{
# Rds default
read_fun <- get_reading_fun()
read_fun

\dontrun{
# fst format
read_fun <- get_reading_fun(ext="fst")
read_fun
}
}
\keyword{internal}
