% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write.R
\name{get_save_fun}
\alias{get_save_fun}
\title{Saving function depending on format selected}
\usage{
get_save_fun(ext = "Rds")
}
\arguments{
\item{ext}{character: format or extension of file. Default is
`fs::path_ext(file)`}
}
\value{
saving function according to `ext`
}
\description{
Saving function depending on format selected
}
\examples{
# Rds default
save_fun <- get_save_fun()
save_fun

# fst format
save_fun <- get_save_fun(ext="fst")
save_fun
}
