% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write.R
\name{get_saving_fun}
\alias{get_saving_fun}
\title{Saving function depending on format selected}
\usage{
get_saving_fun(ext = "Rds")
}
\arguments{
\item{ext}{character: format or extension of file. Default is
\code{fs::path_ext(file)}}
}
\value{
saving function according to \code{ext} that returns logical value
depending on whether the file was saved successfully
}
\description{
Saving function depending on format selected
}
\examples{
# Rds default
save_fun <- get_saving_fun()
save_fun

# fst format
\dontrun{
save_fun <- get_saving_fun(ext="fst")
save_fun
}
}
