% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureLinear.r
\name{mixtureLinear}
\alias{mixtureLinear}
\title{Conduct asociation analysis with linear test under mixture model}
\usage{
mixtureLinear(perm, BetaList, SeList, CorList, ListofMAF, WeightList = NULL)
}
\arguments{
\item{perm}{number of permutations}

\item{BetaList}{list of genetic effect estimates from each study (S vectors)}

\item{SeList}{list of standart errors from each study (S vectors)}

\item{CorList}{list of  LD pairwise correlation between SNPs (S vectors)}

\item{ListofMAF}{list of MAFs (S vectors)}

\item{WeightList}{list of square root of Weights, identity matrix by default (S vectors)}
}
\value{
\code{mixtureQuad} returns a list object with parameters:
\item{pvalue}{p-values}
\item{LogLR}{log likelihood ratio}
\item{Est}{list of estimates}
\item{posterPr}{Posterior probability}
}
\description{
Conduct asociation analysis with linear test under mixture model
}
\examples{
data(Example_SC11)
\donttest{mixtureLinear(1000,Example_SC11$ListBeta,Example_SC11$ListSE,Example_SC11$ListCor,Example_SC11$MafList) }

}
