% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_info}
\alias{path_info}
\title{Get file names and paths}
\usage{
path_info(
  file,
  ext = fs::path_ext(file),
  st_dir = NULL,
  vintage = getOption("stamp.vintage"),
  vintage_dir = NULL,
  recurse = FALSE,
  st_name = NULL
)
}
\arguments{
\item{file}{character: File or connection to write to}

\item{ext}{character: format or extension of file. Default is
\code{fs::path_ext(file)}}

\item{st_dir}{character: Directory to store stamp files. By default it is a
subdirectory at the same level of \code{file}.}

\item{vintage}{logical: Whether to save vintage versions \code{x}. Default \code{TRUE}}

\item{vintage_dir}{character: Directory to save vintages of \code{x}. By default
it is a subdirectory at the same level of \code{file}}

\item{recurse}{logical: is \code{TRUE} if directory in \code{file} does not it will
be created. Default is FALSE}

\item{st_name}{character: name of stamp in file. All stamp files are prefixed
with value in option "stamp.stamp_prefix", which by default is "st_". You
don't need to add the prefix.}
}
\value{
list of directories and files information
}
\description{
Get file names and paths
}
\keyword{internal}
