% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_available}
\alias{pkg_available}
\title{Check whether format is supported and package is available}
\usage{
pkg_available(ext)
}
\arguments{
\item{ext}{character: extension of file}
}
\value{
logical vector for availability of package
}
\description{
Check whether format is supported and package is available
}
\examples{
\dontrun{
pkg_available("fst")
}
}
