% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rand_name}
\alias{rand_name}
\title{General random name}
\usage{
rand_name(l = 8)
}
\arguments{
\item{l}{numeric: length of name. Default 8}
}
\value{
random string name
}
\description{
General random name
}
\keyword{internal}
