% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{st_assert_pk}
\alias{st_assert_pk}
\title{Assert that a data.frame conforms to its schema (if present)}
\usage{
st_assert_pk(x)
}
\description{
Assert that a data.frame conforms to its schema (if present)
}
