% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write.R
\name{st_attr}
\alias{st_attr}
\title{Add attributes and characteristics of x to stamp file}
\usage{
st_attr(
  x,
  hash = NULL,
  complete_stamp = getOption("stamp.completestamp"),
  algo = getOption("stamp.digest.algo")
)
}
\arguments{
\item{x}{R object to write to disk as per limitations of \code{file} format.}

\item{hash}{character: stamp previously calculated. otherwise it will be
added}

\item{complete_stamp}{logical: Whether to add a complete report of data.frame
to stamp file. You need the \code{skimr} package. If \code{skimr} is not in
namespace, limited but lighter report will be added.}

\item{algo}{character: Algorithm to be used in \code{\link[digest:digest]{digest::digest()}}. Default is
"sha1"}
}
\value{
list of attributes
}
\description{
Add attributes and characteristics of x to stamp file
}
\examples{
x <- data.frame(a = 1:10, b = letters[1:10])
st_attr(x)
}
\keyword{internal}
