% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_registry.R
\name{st_formats}
\alias{st_formats}
\title{Inspect available formats
List available registered formats}
\usage{
st_formats()
}
\value{
Character vector of format names.
}
\description{
Return a sorted character vector with the names of formats
currently registered in the internal registry.
}
\examples{
stamp::st_formats()
}
