% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashing.R
\name{st_hash_file}
\alias{st_hash_file}
\title{SipHash-1-3 of a file (bytes on disk)}
\usage{
st_hash_file(path)
}
\arguments{
\item{path}{Path to a file.}
}
\value{
Lowercase hex string (16 hex chars).
}
\description{
SipHash-1-3 of a file (bytes on disk)
}
\keyword{internal}
