% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_store.R
\name{st_lineage}
\alias{st_lineage}
\title{Show immediate or recursive parents for an artifact}
\usage{
st_lineage(path, depth = 1L, alias = NULL)
}
\arguments{
\item{path}{Artifact path (child)}

\item{depth}{Integer depth >= 1. Use Inf to walk recursively.}
}
\value{
data.frame with columns: level, child_path, child_version, parent_path, parent_version
}
\description{
Show immediate or recursive parents for an artifact
}
