% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_store.R
\name{st_load_version}
\alias{st_load_version}
\title{Load a specific version of an artifact}
\usage{
st_load_version(path, version_id, ...)
}
\arguments{
\item{path}{file or directory path}

\item{version_id}{Optional version id of \code{path} to match. Default: any.}

\item{...}{forwarded to format reader}
}
\description{
Load a specific version of an artifact
}
