% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{st_opts_get}
\alias{st_opts_get}
\title{Convenience getter (optional sugar)}
\usage{
st_opts_get(key = NULL)
}
\arguments{
\item{key}{Option name or \code{NULL} for all}
}
\value{
Value or list of all values
}
\description{
Convenience getter (optional sugar)
}
