% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_core.R
\name{st_path}
\alias{st_path}
\title{Declare a path (with optional format)}
\usage{
st_path(path, format = NULL)
}
\arguments{
\item{path}{file or directory path}

\item{format}{optional explicit format ("qs2","rds","csv","fst","json")}
}
\value{
list with class 'st_path'
}
\description{
Declare a path (with optional format)
}
