% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{st_query_table}
\alias{st_query_table}
\title{Query a schema-tagged table by primary-key values (no partitioning)}
\usage{
st_query_table(file, filter, select = NULL, drop_attr = FALSE)
}
\arguments{
\item{file}{artifact path (whole table)}

\item{filter}{named list like list(country="ARG", year=2019, reporting_level=c("national","urban"))}

\item{select}{optional character vector of columns to return (in addition to PK)}

\item{drop_attr}{logical; drop schema attribute on the returned subset}
}
\value{
data.frame subset
}
\description{
Query a schema-tagged table by primary-key values (no partitioning)
}
