% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_read.R
\name{st_read}
\alias{st_read}
\title{Read files with Stamp}
\usage{
st_read(
  file,
  st_dir = NULL,
  vintage = NULL,
  vintage_dir = NULL,
  verbose = getOption("stamp.verbose"),
  ...
)
}
\arguments{
\item{file}{character: file path to be read.}

\item{st_dir}{character: Directory to store stamp files. By default it is a
subdirectory at the same level of \code{file}.}

\item{vintage}{An integer or a quoted directive. "available": displays list
of available versions for \code{measure}. "select"|"pick"|"choose": allows user
to select the vintage of \code{measure}. if the integer is a zero or a negative
number (e.g., \code{-1}), \code{pip_load_aux} will load that number of versions
before the most recent version available. So, if \code{0}, it loads the current
version. If \code{-1}, it will load the version before the current, \code{-2} loads
two versions before the current one, and so on. If it is a positive number,
it must be quoted (as character) and in the form "\%Y\%m\%d\%H\%M\%S". If "00",
it load the most recent version of the data (similar to \code{version = 0} or
\code{version = NULL} or \code{version = "0"}). The difference is that \code{"00"} load
the most recent version of the vintage folder, rather than the current
version in the dynamic folder. Thus, attribute "version" in \code{attr(dd, "version")} is the actual version of the most recent vintage of the file
rather that \code{attr(dd, "version")} equal to "current", which is the default.
Option "00" is useful for vintage control}

\item{vintage_dir}{character: Directory to save vintages of \code{x}. By default
it is a subdirectory at the same level of \code{file}}

\item{verbose}{logical: whether to display additional information. This could
be changed in option \code{"stamp.verbose"}. Default is \code{TRUE}}

\item{...}{not is used right now}
}
\description{
Read files with Stamp
}
