% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{st_set_pk}
\alias{st_set_pk}
\title{Attach a primary-key schema to a data.frame (and return it)}
\usage{
st_set_pk(x, pk, domain = NULL, unique = TRUE)
}
\arguments{
\item{x}{data.frame}

\item{pk}{character vector of column names that form the primary key}

\item{domain}{character scalar or vector; free-form labels for the dataset}

\item{unique}{logical; if TRUE, assert uniqueness of pk combinations}
}
\value{
x with attr(x, "stamp_schema") set
}
\description{
Attach a primary-key schema to a data.frame (and return it)
}
