% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_core.R
\name{st_should_save}
\alias{st_should_save}
\title{Decide if a save should proceed given current st_opts()
Uses versioning policy and code-change rule.}
\usage{
st_should_save(path, x = NULL, code = NULL, alias = NULL)
}
\arguments{
\item{path}{Artifact path on disk.}

\item{x}{Current in-memory object (for content comparison).}

\item{code}{Optional function/expression/character (for code comparison).}

\item{alias}{Optional stamp alias to target a specific stamp folder.}
}
\value{
\verb{list(save = <lgl>, reason = <chr>, latest_version_id = <chr or NA>)}
}
\description{
Decide if a save should proceed given current st_opts()
Uses versioning policy and code-change rule.
}
