% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_store.R
\name{st_versions}
\alias{st_versions}
\title{List versions for an artifact path}
\usage{
st_versions(path)
}
\arguments{
\item{path}{file or directory path}
}
\description{
List versions for an artifact path
}
