% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_registry.R
\name{stamp-format-helpers}
\alias{stamp-format-helpers}
\title{Internal format handlers for stamp}
\arguments{
\item{x}{R object to save.}

\item{path}{Destination file path.}

\item{...}{Additional arguments passed to the underlying writer/reader.}
}
\description{
Internal format handlers for stamp
}
\keyword{internal}
