% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp_clean}
\alias{stamp_clean}
\title{Clean .stamp env}
\usage{
stamp_clean(st_name = NULL, verbose = getOption("stamp.verbose"))
}
\arguments{
\item{st_name}{chracter: stamp name to clean. default is NULL, which cleans
all names}

\item{verbose}{logica: whether to display additional information.}
}
\value{
invisible TRUE is something was clened. FALSE otherwise
}
\description{
Clean .stamp env
}
\examples{
stamp_clean()
}
