% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp_confirm}
\alias{stamp_confirm}
\title{Title}
\usage{
stamp_confirm()
}
\description{
Title
}
\seealso{
Other stamp functions: 
\code{\link{stamp_get}()},
\code{\link{stamp_set}()},
\code{\link{stamp_time}()}
}
\concept{stamp functions}
