% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp_read}
\alias{stamp_read}
\title{Read Stamp in disk}
\usage{
stamp_read(
  st_file = NULL,
  st_dir = NULL,
  st_name = NULL,
  st_ext = getOption("stamp.default.ext"),
  stamp_set = FALSE,
  replace = FALSE
)
}
\arguments{
\item{st_dir}{character: parent directory to store stamp file (see details).}

\item{st_name}{character: name of stamp in file. All stamp files are prefixed
with value in option "stamp.stamp_prefix", which by default is "st_". You
don't need to add the prefix.}

\item{st_ext}{character: format of stamp file to save. Default is value in
option "stamp.default.ext"}

\item{stamp_set}{logical: whether to set stamp in .stamp env, using
\code{st_name}.}

\item{replace}{Logical: if TRUE and \code{st_name} already exists in \code{.stamp}
environment, it will be replaced with new stamp. If \code{FALSE} it gives an
error. Default is \code{FALSE}}
}
\value{
stamp list invisibly
}
\description{
Read Stamp in disk
}
\examples{
\dontrun{

x <- data.frame(a = 1:5,
b = letters[1:5])

st_dir <- tempdir()
st_name <- "xst"
sv <- stamp_save(x = x,
st_dir = st_dir,
st_name = st_name)

nsv <- names(sv) |>
 fs::path()

stamp_read(nsv)

}
}
\seealso{
Other stamp functions: 
\code{\link{stamp_confirm}()},
\code{\link{stamp_get}()},
\code{\link{stamp_save}()},
\code{\link{stamp_set}()},
\code{\link{stamp_time}()},
\code{\link{stamp_x_attr}()}
}
\concept{stamp functions}
