% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stamp.R
\name{stamp_read_defense}
\alias{stamp_read_defense}
\title{defenses of stamp_read}
\usage{
stamp_read_defense(
  st_file = NULL,
  st_dir = NULL,
  st_name = NULL,
  st_ext = getOption("stamp.default.ext"),
  stamp_set = FALSE,
  replace = FALSE,
  verbose = getOption("stamp.verbose")
)
}
\arguments{
\item{st_file}{character: file path to be read}

\item{st_dir}{character: parent directory to store stamp file (see details).}

\item{st_name}{character: name of stamp in file. All stamp files are prefixed
with value in option "stamp.stamp_prefix", which by default is "st_". You
don't need to add the prefix.}

\item{st_ext}{character: format of stamp file to save. Default is value in
option "stamp.default.ext"}

\item{stamp_set}{logical: whether to set stamp in .stamp env, using
\code{st_name}.}

\item{replace}{Logical: if TRUE and \code{st_name} already exists in \code{.stamp}
environment, it will be replaced with new stamp. If \code{FALSE} it gives an
error. Default is \code{FALSE}}

\item{verbose}{logical: Fi TRUE displays information about stamping process.}
}
\value{
Nothing
}
\description{
defenses of stamp_read
}
\keyword{internal}
