% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp_save}
\alias{stamp_save}
\title{Save Stamp in disk}
\usage{
stamp_save(
  x = NULL,
  st_dir = NULL,
  st_name = NULL,
  st_ext = getOption("stamp.default.ext"),
  stamp = NULL,
  stamp_set = FALSE,
  replace = FALSE,
  x_attr = FALSE,
  verbose = getOption("stamp.verbose"),
  ...
)
}
\arguments{
\item{x}{R object to stamp}

\item{st_dir}{character: parent directory to store stamp file (see details).}

\item{st_name}{character: name of stamp in file. All stamp files are prefixed
with value in option "stamp.stamp_prefix", which by default is "st_". You
don't need to add the prefix.}

\item{st_ext}{character: format of stamp file to save. Default is value in
option "stamp.default.ext"}

\item{stamp}{list of stamp from stamp_get() in case it was calculated before
hand. Developers option. It should be used interactively.}

\item{stamp_set}{logical: whether to set stamp in .stamp env, using
\code{st_name}.}

\item{replace}{Logical: if TRUE and \code{st_name} already exists in \code{.stamp}
environment, it will be replaced with new stamp. If \code{FALSE} it gives an
error. Default is \code{FALSE}}

\item{x_attr}{logical: whether or not to save the attributes of \code{x} along
with the stamp. Useful for quick comparisons. Default is FALSE}

\item{verbose}{logical: Fi TRUE displays information about stamping process.}

\item{...}{other arguments passed to stamp_get()}
}
\value{
TRUE is saved correctly. FALSE otherwise
}
\description{
Create and save in file stamp for future use
}
\details{
\code{st_dir} is parent directory. It is inside \code{st_dir} that {stamp}
creates another subdirectory with name in option "stamp.dir_stamp" and it
is in there where the stamps are saved. The idea objective is to have a
directory for stamps only. By default, \code{st_dir} is the current directory.
If last directory name of \code{st_dir} is equal to option "stamp.dir_stamp",
then \code{st_dir} becomes the stamps directory.

\code{st_name} must prefixed to avoid overwriting actual data. This is just a
precaution that should not present bumps in any workflow.  If the beginning
of \code{st_name} is identical to the value in "stamp.stamp_prefix", then it is
adopted as is. Otherwise, the prefix in "stamp.stamp_prefix" will be added
to \code{st_name}. If NULL, \code{st_name} would be a random name of 8 characters.
}
\examples{
\dontrun{

x <- data.frame(a = 1:5,
b = letters[1:5])

st_dir <- tempdir()
st_name <- "xst"
stamp_save(x = x,
st_dir = st_dir,
st_name = st_name)

}
}
\seealso{
Other stamp functions: 
\code{\link{stamp_confirm}()},
\code{\link{stamp_get}()},
\code{\link{stamp_read}()},
\code{\link{stamp_set}()},
\code{\link{stamp_time}()},
\code{\link{stamp_x_attr}()}
}
\concept{stamp functions}
