% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp_x_attr}
\alias{stamp_x_attr}
\title{Add attributes and characteristics of x to be used in stamp file}
\usage{
stamp_x_attr(x, complete_stamp = getOption("stamp.completestamp"))
}
\arguments{
\item{x}{R object to write to disk as per limitations of \code{file} format.}

\item{complete_stamp}{logical: Whether to add a complete report of data.frame
to stamp file. You need the \code{skimr} package. If \code{skimr} is not in
namespace, limited but lighter report will be added.}
}
\value{
list of attributes
}
\description{
Add attributes and characteristics of x to be used in stamp file
}
\examples{
x <- data.frame(a = 1:10, b = letters[1:10])
stamp_x_attr(x)
}
\seealso{
Other stamp functions: 
\code{\link{stamp_confirm}()},
\code{\link{stamp_get}()},
\code{\link{stamp_save}()},
\code{\link{stamp_set}()},
\code{\link{stamp_time}()}
}
\concept{stamp functions}
