% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage.R
\name{vt_compare}
\alias{vt_compare}
\title{compare changes vintage of R objects}
\usage{
vt_compare(waldo = getOption("stamp.waldo"))
}
\arguments{
\item{waldo}{logical: whether to display waldo::compare results if objects
are different. Requires package waldo. If not in namespace,
base::identical() will be executed.}
}
\value{
Logic value. \code{TRUE} if objects are identival, \code{FALSE} otherwise.
}
\description{
Compares where the differences are between R objects using waldo
package. You can select whether to compare vintages with the R object in used
or between any of the vintages files saved
}
