% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write.R
\name{write}
\alias{write}
\title{Save objects in disk according to desired file}
\usage{
write(x, lp)
}
\arguments{
\item{x}{R object to write to disk as per limitations of \code{file} format.}

\item{lp}{list of paths returned by \code{path_info()}}
}
\value{
logical value if it was waved correctly
}
\description{
Save objects in disk according to desired file
}
\keyword{internal}
