% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write.R
\name{check_complex_data}
\alias{check_complex_data}
\title{Check whther object can be saved in tabular formats like fst}
\usage{
check_complex_data(x)
}
\arguments{
\item{x}{R object to write to disk as per limitations of \code{file} format.}
}
\value{
logical for complex data
}
\description{
Check whther object can be saved in tabular formats like fst
}
\examples{
# False
check_complex_data(data.frame())

# TRUE
check_complex_data(list())
}
