% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write.R
\name{deal_with_file_path}
\alias{deal_with_file_path}
\title{Make sure file names and directory paths are working fine}
\usage{
deal_with_file_path(file, ext, create_dir)
}
\arguments{
\item{file}{character: File or connection to write to}

\item{ext}{character: format or extension of file. Default is
\code{fs::path_ext(file)}}

\item{create_dir}{logical: is \code{TRUE} if directory in \code{file} does not it will
be created. Default is FALSE}
}
\value{
character vector with file path
}
\description{
Make sure file names and directory paths are working fine
}
\keyword{internal}
