% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{.st_apply_retention}
\alias{.st_apply_retention}
\title{Optionally invoked after st_save() to apply retention for a single artifact}
\usage{
.st_apply_retention(artifact_path)
}
\description{
Optionally invoked after st_save() to apply retention for a single artifact
}
\keyword{internal}
