% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.st_artifact_path}
\alias{.st_artifact_path}
\title{Compute the actual artifact path in .st_data structure (internal)}
\usage{
.st_artifact_path(rel_path, alias = NULL)
}
\arguments{
\item{rel_path}{Character relative path from alias root}

\item{alias}{Optional alias}
}
\value{
Character scalar absolute path to the artifact file
}
\description{
Returns the path where the actual user file will be stored.
}
\details{
Structure: <file_storage_dir>/\if{html}{\out{<filename>}}
}
\keyword{internal}
