% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{.st_delete_version_dir_safe}
\alias{.st_delete_version_dir_safe}
\title{Safely delete a version directory (internal)}
\usage{
.st_delete_version_dir_safe(vdir)
}
\description{
Safely delete a version directory (internal)
}
\keyword{internal}
