% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.st_extract_rel_path}
\alias{.st_extract_rel_path}
\title{Extract relative path from an absolute path (internal)}
\usage{
.st_extract_rel_path(abs_path, alias = NULL)
}
\arguments{
\item{abs_path}{Character absolute path}

\item{alias}{Optional alias}
}
\value{
Character relative path from root, or NULL if path not under root/data_folder
}
\description{
Given an absolute path (possibly in .st_data or possibly user's original path),
extract the relative path component from alias root.
}
\details{
This handles both cases:
\itemize{
\item Path is in .st_data: extract rel_path from .st_data structure
\item Path is under root directly: extract rel_path from root
}
}
\keyword{internal}
