% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_store.R
\name{.st_state_dir_abs}
\alias{.st_state_dir_abs}
\title{Absolute state directory path (internal)}
\usage{
.st_state_dir_abs(alias = NULL)
}
\value{
Character scalar absolute path to the state directory.
}
\description{
Compute the absolute path to the package state directory. This is
constructed as \if{html}{\out{<root>}}/<state_dir> where \code{root} is from
\code{st_init()} and \code{state_dir} is an option stored in the package state.
}
\keyword{internal}
