% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_st_name}
\alias{format_st_name}
\title{Format Stamp name}
\usage{
format_st_name(
  st_name = NULL,
  st_nm_pr = getOption("stamp.stamp_prefix"),
  seed = NULL
)
}
\arguments{
\item{st_name}{character: name of stamp in file. All stamp files are prefixed
with value in option "stamp.stamp_prefix", which by default is "st_".
You don't need to add the prefix.}

\item{st_nm_pr}{character: \code{st_name} prefix to save in file. default is value
in option "stamp.stamp_prefix".}

\item{seed}{numeric: seed for random name. Default is NULL so each time the
the random name generated is the same. Use only for replicability purposes}
}
\value{
formatted directory
}
\description{
Format Stamp name
}
\keyword{internal}
