% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_st_name}
\alias{format_st_name}
\title{Format Stamp name}
\usage{
format_st_name(st_name = NULL)
}
\arguments{
\item{st_name}{character: name of stamp in file. All stamp files are prefixed
with value in option "stamp.stamp_prefix", which by default is "\emph{st}".}
}
\value{
formatted directory
}
\description{
Format Stamp name
}
\keyword{internal}
