% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stamp.R
\name{rand_name}
\alias{rand_name}
\title{General random name}
\usage{
rand_name(seed = NULL, l = 8)
}
\arguments{
\item{seed}{numeric: seed for random name. Default is NULL so each time the
the random name generated is the same. Use only for replicability purposes}

\item{l}{numeric: length of name. Default 8}
}
\value{
random string name
}
\description{
General random name
}
\keyword{internal}
