% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_core.R
\name{st_init}
\alias{st_init}
\title{Initialize stamp project structure}
\usage{
st_init(root = ".", state_dir = ".stamp")
}
\arguments{
\item{root}{project root (default ".")}

\item{state_dir}{directory name for internal state (default ".stamp")}
}
\value{
(invisibly) the absolute state dir
}
\description{
Initialize stamp project structure
}
