% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_store.R
\name{st_latest}
\alias{st_latest}
\title{Get the latest version_id for an artifact path}
\usage{
st_latest(path, alias = NULL)
}
\arguments{
\item{path}{file or directory path}

\item{alias}{Optional stamp alias to target a specific stamp folder.}
}
\description{
Get the latest version_id for an artifact path
}
