% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{st_list_parts}
\alias{st_list_parts}
\title{List available partitions under a base directory}
\usage{
st_list_parts(base, filter = NULL, recursive = TRUE)
}
\arguments{
\item{base}{Base dir}

\item{filter}{Named list to restrict partitions (exact match on key fields)}

\item{recursive}{Logical; search subdirs (default TRUE)}
}
\value{
A data.frame with columns: path plus one column per partition key
}
\description{
List available partitions under a base directory
}
