% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{st_load_parts}
\alias{st_load_parts}
\title{Load and row-bind partitioned data}
\usage{
st_load_parts(base, filter = NULL, as = c("rbind", "dt"))
}
\arguments{
\item{base}{Base dir}

\item{filter}{Named list to restrict partitions (exact match)}

\item{as}{Data frame binding mode: "rbind" (base) or "dt" (data.table if available)}
}
\value{
Data frame with unioned columns and extra columns for the key fields
}
\description{
Load and row-bind partitioned data
}
