% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_write.R
\name{st_write}
\alias{st_write}
\title{Write R object with corresponding stamp (hash digest)}
\usage{
st_write(
  x,
  file,
  ext = fs::path_ext(file),
  st_dir = NULL,
  attr = list(),
  create_dir = FALSE,
  force = FALSE,
  algo = "sha1",
  vintage = TRUE,
  vintage_dir = NULL,
  ...
)
}
\arguments{
\item{x}{R object to write to disk as per limitations of `file` format.}

\item{file}{character: File or connection to write to}

\item{ext}{character: format or extension of file. Default is
`fs::path_ext(file)`}

\item{st_dir}{character: Directory to store stamp files. By default it is a
subdirectory at the same level of `file`.}

\item{attr}{list of attributes to store}

\item{create_dir}{logical: is `TRUE` if directory in `file` does not it will
be created. Default is FALSE}

\item{force}{logical: replace file in disk even if has hasn't changed}

\item{algo}{character: Algorithm to be used in [digest::digest()]. Default is
"sha1"}

\item{vintage}{logical: Whether to save vintage versions `x`. Default `TRUE`}

\item{vintage_dir}{character: Directory to save vintages of `x`. By default
it is a subdirectory at the same level of `file`}

\item{...}{not is used right now}
}
\value{
TRUE is object was saved successfully. FALSE otherwise.
}
\description{
Write R object with corresponding stamp (hash digest)
}
\section{Details}{
 Object `x` is stored in `file` but its hash (i.e., stamp)
  is stored in subdirectory `st_file`.
}

\examples{
\dontrun{
  tfile <- file_temp(ext = "qs")
  st_write(df, tfile)
}
}
