% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-package.R
\docType{package}
\name{stamp-package}
\alias{stamp}
\alias{stamp-package}
\title{stamp: Read and Write R Objects with Data Hash}
\description{
A system to manage the reading and writing R objects in disk by creating and saving data signatures (hash).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/randrescastaneda/stamp}
  \item \url{https://randrescastaneda.github.io/stamp/}
  \item Report bugs at \url{https://github.com/randrescastaneda/stamp/issues}
}

}
\author{
\strong{Maintainer}: R.Andres Castaneda \email{acastanedaa@worldbank.org} (\href{https://orcid.org/0000-0003-1154-1374}{ORCID})

}
\keyword{internal}
