% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp-package.R
\docType{package}
\name{stamp-package}
\alias{stamp}
\alias{stamp-package}
\title{stamp: Reproducible Artifact Store with Sidecars, Versions, and Partitions}
\description{
Stamp saves R objects with sidecar metadata (hashes, provenance, PKs), keeps version history, prunes old versions, and supports Hive-style partitions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://randrescastaneda.github.io/stamp/}
  \item \url{https://github.com/randrescastaneda/stamp}
  \item Report bugs at \url{https://github.com/randrescastaneda/stamp/issues}
}

}
\author{
\strong{Maintainer}: R.Andres Castaneda \email{acastanedaa@worldbank.org} (\href{https://orcid.org/0000-0003-1154-1374}{ORCID})

Authors:
\itemize{
  \item Diana C. Garcia Rojas \email{dgarciarojas@worldbank.org} (\href{https://orcid.org/0000-0001-9423-3865}{ORCID})
}

}
\keyword{internal}
