% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp_confirm}
\alias{stamp_confirm}
\title{Confirm stamp has not changed}
\usage{
stamp_confirm(
  x,
  verbose = getOption("stamp.verbose"),
  using = c("self", "stamp"),
  ...
)
}
\arguments{
\item{x}{An arbitrary R object which will then be passed to the
base::serialize function}

\item{verbose}{logical: whether to display additional information. This could
be changed in option \code{"stamp.verbose"}. Default is \code{TRUE}}

\item{...}{not is used right now}
}
\value{
Logical value. \code{FALSE} if the objects do not match and  \code{TRUE} if
they do.
}
\description{
verifies that, were the stamp recalculated, it would match the
one previously set with stamp_set().
}
\seealso{
Other stamp functions: 
\code{\link{stamp_attr}()},
\code{\link{stamp_get}()},
\code{\link{stamp_set}()},
\code{\link{stamp_time}()}
}
\concept{stamp functions}
