% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp_time}
\alias{stamp_time}
\title{Get time parameters}
\usage{
stamp_time()
}
\value{
list of time parameters as objects
}
\description{
It uses the values stored in "stamp.timezone", "stamp.timeformat" and
"stamp.usetz" options
}
\examples{
stamp_time()
}
\seealso{
Other stamp functions: 
\code{\link{stamp_attr}()},
\code{\link{stamp_confirm}()},
\code{\link{stamp_get}()},
\code{\link{stamp_save}()},
\code{\link{stamp_set}()}
}
\concept{stamp functions}
