% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_write.R
\name{write}
\alias{write}
\title{Save objects in disk according to desired file}
\usage{
write(x, file, ext = fs::path_ext(file))
}
\arguments{
\item{x}{R object to write to disk as per limitations of `file` format.}

\item{file}{character: File or connection to write to}

\item{ext}{character: format or extension of file. Default is
`fs::path_ext(file)`}
}
\value{
logical value if it was waved correctly
}
\description{
Save objects in disk according to desired file
}
\keyword{internal}
